rename Solver and SolverGui structs to reflect their uses
