+++ src/ui/solver.rs
